feat: added unit tests and custom validators
It is now possible to add a customValidator. This is an optional
field that holds a user function that takes the entire object, and
must return a boolean.

The custom validators are fired after the data has been validated
and confirmed to be of the schema's type, it can thus also be used
to validate the data based on other field values.

Additionally the repo now also includes unit tests. The reason for
adding unit tests is self-explainatory.